requirements.txt
pytesseract
pdf2image
Pillow
streamlit
